Worked on playerCharacter
Items implemented and appear to be working properly.

Is there anything I need to do for character creation, or is that being
done entirely in the GUI? Because I don't know how the GUI will tie in.

Also, if the ö in Gödel doesn't work properly, we can turn it into an o.
Just let me know.